s19

+++ service/todo.go
@@ -273,6 +273,10 @@ func (s *TODOService) UpdateTODO(ctx context.Context, id int64, subject, descrip
 func (s *TODOService) DeleteTODO(ctx context.Context, ids []int64) error {
 	const deleteFmt = `DELETE FROM todos WHERE id IN (?%s)`
 
+	if len(ids) == 0 {
+		return nil
+	}
+
 	symbols := strings.Repeat(", ?", len(ids)-1)
 	delete := fmt.Sprintf("DELETE FROM todos WHERE id IN (?%s)", symbols)
